Moved away from prototype component and toward Routing
