+++ frontend/src/App.jsx
@@ -1,19 +1,23 @@
 import { ChakraProvider } from "@chakra-ui/react";
-import DemoPage from "./routes/DemoPage";
+import HomePage from "./routes/HomePage";
+import ErrorPage from "./routes/ErrorPage";
 import { RouterProvider, createBrowserRouter } from "react-router-dom";
 
 
 const router = createBrowserRouter([
 	{
 		path: "/",
-		element: <DemoPage />
+		element: <HomePage />,
+		errorElement: <ErrorPage />
 	}
 ])
 
 function App() {
-	return <ChakraProvider>
-		<RouterProvider router={router} />
-	</ChakraProvider>;
+	return (
+		<ChakraProvider>
+			<RouterProvider router={router} />
+		</ChakraProvider>
+	);
 }
 
 export default App;
