+++ frontend/src/components/JobResults/ResultContainer.jsx
@@ -3,16 +3,20 @@ import JobCard from "./JobCard";
 
 
 const ResultContainer = ({ jobResults }) => {
-	return <Wrap>
-		{jobResults.length > 0 ?
-			jobResults.map((result) => {
-				return <WrapItem key={result.job_id}>
-					<JobCard jobData={result} />
-				</WrapItem>
-			})
-			: ""
-		}
-	</Wrap>
+	return (
+		<Wrap>
+			{jobResults.length > 0 ?
+				jobResults.map((result) => {
+					return (
+						<WrapItem key={result.job_id}>
+							<JobCard jobData={result} />
+						</WrapItem>
+					);
+				})
+				: ""
+			}
+		</Wrap>
+	);
 }
 
 export default ResultContainer;
