+++ frontend/src/components/NavHeader.jsx
@@ -1,0 +1,8 @@
+const NavHeaders = () => {
+	return (
+		<>
+		</>
+	);
+}
+
+export default NavHeaders;
