+++ frontend/src/routes/HomePage.jsx
@@ -1,0 +1,10 @@
+import Dashboard from "../components/Dashboard"
+
+// TODO: Should hold the Header, Dashboard, and Footer components
+const HomePage = () => {
+	return (
+		<Dashboard />
+	);
+}
+
+export default HomePage;
